requirements.txt
pypdf
pinecone
python-dotenv
google-generativeai